yessir

+++ db/migrate/20230404191936_change_price_datatype_in_products.rb
@@ -1,5 +1,5 @@
 class ChangePriceDatatypeInProducts < ActiveRecord::Migration[7.0]
   def change
-    change_column :products, :price, :decimal
+    change_column :products, :price, :float
   end
 end
